movie_review_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tortoise-orm[asyncpg]==0.20.0
aerich==0.7.2
python-multipart==0.0.6
numpy==1.26.2 